requirements.txt
# 大豐環保人力變化視覺化系統 - 套件需求
# Python 3.9+

# Web 框架
streamlit>=1.28.0

# 資料處理
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0           # 支援舊版 .xls 檔案

# 圖表
plotly>=5.18.0

# 設定檔
pyyaml>=6.0

# 日期處理
python-dateutil>=2.8.0

# Cloudflare R2 上傳（S3 相容）
boto3>=1.28.0
python-dotenv>=1.0.0

# Email（標準庫，無需安裝）
# smtplib - 內建
# email - 內建
